math-mentor/requirements.txt
streamlit
python-dotenv
openai
faiss-cpu
numpy
pandas
pillow
